socket notification on chat_message
only the process "owning" the websocket can send messages. So we have to
replace our socket-conn (registry) with a subject. This subject can be
populated with our messages (from other pids). Inside mist.socket the
subject is converted into a select which is returned to mist and the
message type comes out in the handler under Custom(_). Makes kind of sense once
you think about it - no poll can happen bc our async runtime can't
mutate - so we can only use the one 'channel' we have.