add confirmation propagation (tries socket if available, fallback to http)
confirmation of the confirmation - so confirmation ^2 goes back to original sender (on the way he communicated so either in http-response or via another socket message).
This makes it ideomatic to just update the delivery-status upon
receiving the confirmation-message on client side. The downside is that
the ambiguity of sending this shite around is obfuscating the flow of
communication a bit. there's a excalidraw in /docs trying to explain
some of my thoughts...